+++ app/smartbrain_3000/web/_app/immutable/entry/start.C_MLvKgo.js
@@ -1,1 +1,0 @@
-import{a as e,n as t}from"../chunks/DA8se2FU.js";export{e as load_css,t as start};